:wrench: Allow loading from disk
