+++ src/renderer/themes.ts
@@ -1,62 +1,34 @@
-import * as MonacoType from 'monaco-editor';
+import * as fsType from 'fs-extra';
+import * as path from 'path';
 
-export interface FiddleTheme {
-  editor: Partial<MonacoType.editor.IStandaloneThemeData>;
-  fonts: {
-    common: string;
-  };
-  colors: {
-    'foreground-brighter': string;
-    'foreground-bright': string;
-    'foreground-tuned': string;
-    'background-3': string;
-    'background-2': string;
-    'background-1': string;
-    'border-color-dark': string;
-    'border-color': string;
-    border: string;
-    'text-color': string;
-    'text-color-dark': string;
-    'text-color-bright': string;
-    'error-color': string;
-  };
-  css: string;
-}
+import { CONFIG_PATH } from '../constants';
+import { fancyImport } from '../utils/import';
+import { defaultDark, DefaultThemes, FiddleTheme } from './themes-defaults';
 
 /**
- * The Monaco editor theme used by Electron Fiddle.
+ * Returns a Fiddle theme, either a default one or by checking
+ * the disk for a JSON file.
+ *
+ * @export
+ * @param {string} [name]
+ * @returns {Promise<FiddleTheme>}
  */
-export const defaultDark = {
-  common: {
-    'foreground-brighter': '#9feafa',
-    'foreground-bright': '#8ac7d6',
-    'foreground-tuned': '#608291',
-    'background-3': '#2c2e3b',
-    'background-2': '#1d2427',
-    'background-1': '#2f3241',
-    'border-color-dark': '#1e2527',
-    'border-color': '#5c5f71',
-    border: '1px solid var(--border-color)',
-    'text-color': '#d4d4d4',
-    'text-color-dark': '#1e2527',
-    'text-color-bright': '#dcdcdc',
-    'error-color': '#df3434',
-      // tslint:disable-next-line:max-line-length
-    'fonts-common': `-apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";`
-  },
-  editor: {
-    base: 'vs-dark',
-    inherit: true,
-    rules: [{ background: '2f3241' }],
-    colors: {
-      'editor.background': '#2f3241'
+export async function getTheme(name?: string): Promise<FiddleTheme> {
+  let theme = !name || name === DefaultThemes.DARK
+    ? defaultDark
+    : defaultDark;
+  let cssContent = '';
+
+  if (name) {
+    const fs = await fancyImport<typeof fsType>('fs-extra');
+    const themePath = path.join(CONFIG_PATH, 'themes', `${name}.json`);
+
+    try {
+      theme = await fs.readJSON(themePath);
+    } catch (error) {
+      console.warn(`Themes: Loading theme ${name} failed`, error);
     }
   }
-};
-
-export async function getTheme(): Promise<FiddleTheme> {
-  const theme = defaultDark;
-  let cssContent = '';
 
   Object.keys(theme.common).forEach((key) => {
     cssContent += `  --${key}: ${theme.common[key]};\n`;
